minor project complete

+++ minor-1/src/main/java/com/example/minor_1/models/Transaction.java
@@ -34,7 +34,7 @@ public class Transaction {
     @UpdateTimestamp
     private Date updatedOn;
 
-    private Double fine;
+    private Integer fine;
 
     @ManyToOne
     @JoinColumn
